+++ styles/TeamScreen/TeamScreen.js
